SegFormer_IDD/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
datasets>=2.16.0
huggingface-hub>=0.17.0
accelerate>=0.25.0
numpy>=1.24.0
Pillow>=9.5.0
tqdm>=4.65.0
matplotlib>=3.7.2
scikit-learn>=1.3.0
opencv-python>=4.8.0
